Modificado o algoritmo base  e adicionado a opcao de importar o texto apartir de um arquivo

+++ src/àrvore/Folha.java
@@ -13,7 +13,7 @@ public class Folha extends Arvore{
     
     public char letras;
     
-    public Folha(int Frequencia , char Letras){
+    public Folha(long Frequencia , char Letras){
         super(Frequencia);
         this.letras = Letras;
     }
